IDE-2909 Add validation message when there is no files need to be upgraded
